analysis: resolve actionable findings
Fixes: 2b7b06719
Fixes: grommunio-web-1.0-15-gcffb8040e
Fixes: grommunio-web-1.0-22-g7fe0804e5
Fixes: grommunio-web-1.0-42-g2640d47d9
Fixes: grommunio-web-1.0-133-gc2c543e52
Fixes: grommunio-web-1.0-171-gbf45991cb
Fixes: grommunio-web-3.0-5-gde01c3141
Fixes: grommunio-web-3.1-16-g56545243c
Fixes: grommunio-web-3.1-204-gae1b0f139
Fixes: grommunio-web-3.2-33-g63ef83829
Fixes: grommunio-web-3.7-16-gf014b1ecc
Fixes: grommunio-web-3.9-14-gc74abd181
Fixes: grommunio-web-3.10-27-g306fb9935
Fixes: grommunio-web-3.10-67-g5393c9b25
Fixes: grommunio-web-3.11-21-g404471d27
Fixes: grommunio-web-3.11-46-g952fbb671
Fixes: grommunio-web-3.14-50-g44c4891c7
Fixes: grommunio-web-3.14-61-g52e05da68
Fixes: grommunio-web-3.15-4-g1e1793cea
Fixes: grommunio-web-3.15-49-g624c2c424
Fixes: grommunio-web-3.17-222-g44ca8d08a
Fixes: grommunio-web-3.17-270-g89ec8293b
Fixes: grommunio-web-3.18-15-gae30561ec
Fixes: grommunio-web-3.18-35-g272493132
Fixes: grommunio-web-3.18-42-g643ad2312
Fixes: grommunio-web-3.18-52-g32b910473
Fixes: grommunio-web-3.18-128-gaa627970a
Fixes: grommunio-web-3.19-47-g67475b6f3
Fixes: grommunio-web-3.19-48-g99833b9e1
Fixes: grommunio-web-3.19-124-g12ddaf56b
Fixes: grommunio-web-3.19-282-g2d1dae4d8
Fixes: grommunio-web-3.19-353-g40aed85b3
Fixes: grommunio-web-3.19-362-gef1e91fe0

+++ plugins/files/php/Files/Core/Util/class.stringutil.php
@@ -55,7 +55,7 @@ class StringUtil {
 		// $chars - all allowed characters
 		$chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890";
 
-		mt_srand((float) microtime() * 1000000);
+		mt_srand((int) ((float) microtime() * 1000000));
 		$i = 0;
 		$pass = "";
 		while ($i < $length) {
